Refactor date time validator
