+++ src/test/java/com/networknt/schema/Issue784Test.java
@@ -2,6 +2,8 @@ package com.networknt.schema;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.networknt.schema.SpecVersion.VersionFlag;
+
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
@@ -61,11 +63,14 @@ public class Issue784Test {
     }
 
     private JsonSchema createSchema(boolean useCustomDateFormat) {
-        JsonMetaSchema overrideDateTimeValidator = new JsonMetaSchema
-                .Builder(JsonMetaSchema.getV7().getUri())
-                .idKeyword("$id")
-                .addKeywords(ValidatorTypeCode.getNonFormatKeywords(SpecVersion.VersionFlag.V7))
-                .addFormats(useCustomDateFormat ? Collections.singletonList(new CustomDateTimeFormat()) : Collections.emptyList())
+        JsonMetaSchema overrideDateTimeValidator =JsonMetaSchema
+                .builder(JsonMetaSchema.getV7().getUri(), JsonMetaSchema.getV7())
+                .formats(formats -> {
+                    if (useCustomDateFormat) {
+                        CustomDateTimeFormat format = new CustomDateTimeFormat();
+                        formats.put(format.getName(), format);
+                    }
+                })
                 .build();
 
         return new JsonSchemaFactory
